requirements.txt
numpy==1.16.2
torch==1.0.1.post2
gensim==3.7.1
glob2==0.6
graphviz==0.10.1
matplotlib==3.0.3
nltk==3.4
pandas==0.24.2
pymongo==3.7.2
rouge==0.3.1
seaborn==0.9.0
sentencepiece==0.1.81
scikit_learn==0.20.3
tabulate==0.8.3
tqdm==4.31.1
umap==0.1.1
visdom==0.1.8.8
PyYAML==5.1
